Separated some connection service methods into the standard/consumption files
