+++ libs/designer/src/lib/core/actions/bjsworkflow/add.ts
@@ -181,7 +181,7 @@ export const trySetDefaultConnectionForNode = async (
 ) => {
   const connections = await getConnectionsForConnector(connectorId);
   if (connections.length > 0) {
-    await ConnectionService().createConnectionAclIfNeeded(connections[0]);
+    await ConnectionService().setupConnectionIfNeeded(connections[0]);
     dispatch(updateNodeConnection({ nodeId, connectionId: connections[0].id, connectorId }));
   } else if (isConnectionRequired) {
     dispatch(initEmptyConnectionMap(nodeId));
